+++ app/src/main/java/dev/kostromdan/mods/crash_assistant/app/CrashAssistantApp.java
@@ -62,6 +62,9 @@ public class CrashAssistantApp {
 
     @NoJexl
     public static void main(String[] args) {
+        PlatformHelp.platform=PlatformHelp.FORGE;
+//        PlatformHelp.platform=PlatformHelp.FABRIC;
+//        PlatformHelp.platform=PlatformHelp.NEOFORGE;
         Thread.setDefaultUncaughtExceptionHandler((thread, throwable) -> {
             LOGGER.error("Uncaught exception in \"{}\" thread:", thread.getName(), throwable);
         });